requirements.txt
streamlit>=1.37
pandas
folium
streamlit-folium
python-dotenv
requests
selenium
langchain-core
langchain-ollama